Begin to fix UI tests for time entries page
