+++ tests/login.test.js
@@ -1,48 +1,10 @@
 import { expect, test } from "@playwright/test";
 import LoginPage from "./TestHelper/LoginPage";
 import TroiApiStub, { correctUser, correctPassword } from "./TestHelper/TroiApiStub";
-
-let apiStub;
+import { initilaizeTestSetup } from "./TestHelper/TestHelper";
 
 test.beforeEach(async ({ context }) => {
-    apiStub = new TroiApiStub();
-
-    await context.route(
-        "https://digitalservice.troi.software/api/v2/rest/**",
-        async (route) => {
-            const authnHeader = await route.request().headerValue("Authorization");
-            if (!apiStub.isAuthorized(authnHeader)) {
-                route.fulfill(apiStub.unauthorizedResponse());
-                return;
-            }
-
-            const method = route.request().method();
-            const { pathname, searchParams: params } = new URL(route.request().url());
-            const postData = route.request().postDataJSON();
-            const matchedResponse = apiStub.match(method, pathname, params, postData);
-
-            if (matchedResponse !== null) {
-                route.fulfill(matchedResponse);
-            } else {
-                console.log({ matchedResponse, method, pathname, params, postData });
-                route.abort();
-            }
-        }
-    );
-
-    await context.route("/time_entries/*", async (route) => {
-        const { pathname } = new URL(route.request().url());
-        const method = route.request().method();
-
-        const id = parseInt(pathname.split("/").at(-1));
-
-        if (method === "DELETE") {
-            apiStub.deleteEntry(id);
-            route.fulfill(apiStub._response({}));
-        } else {
-            route.continue();
-        }
-    });
+    initilaizeTestSetup(context);
 });
 
 test.describe("Auth", async () => {
